backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.1
pandas>=1.3.0
numpy>=1.21.0
fastapi>=0.68.0
python-multipart>=0.0.5
uvicorn>=0.15.0
